♻️ Remove unnecessary error handling
Previously the app was unnecessarily accounting for errors which should never really happen in the real world, such as when a new todo being added has the same ID as an existing one. This commit will remove this error handling, and simply crash the app when these unexpected errors occur.